data_ingestion/voice_agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
gTTS==2.3.2
openai-whisper==20231117
python-multipart==0.0.6
pydantic==2.4.2 